+++ cfgov/unprocessed/js/molecules/Notification.js
@@ -1,8 +1,7 @@
 'use strict';
 
-// Required polyfills for <IE9.
-require( '../modules/polyfill/query-selector' );
-require( '../modules/polyfill/class-list' );
+// Required polyfills for IE9.
+if ( !Modernizr.classlist ) { require( '../modules/polyfill/class-list' ); } // eslint-disable-line no-undef, global-require, no-inline-comments, max-len
 
 // Required modules.
 var atomicCheckers = require( '../modules/util/atomic-checkers' );